finyk-domain(pkg): PR2 — extract storage keys + backup normalization
Pull the DOM-free pieces of `apps/web/src/modules/finyk/lib/*` into
`@sergeant/finyk-domain` so the upcoming mobile twin can reuse them
without touching localStorage:

- `src/storageKeys.ts` — `FINYK_STORAGE_KEYS` (transactions /
  categories / budget), `FINYK_MANUAL_ONLY_KEY` (skip-Monobank flag),
  and `FINYK_BACKUP_STORAGE_KEYS` (the 14 backup fields ↔ legacy
  `finyk_*` key names).
- `src/backup.ts` — `FINYK_BACKUP_VERSION`,
  `DEFAULT_FINYK_MONTHLY_PLAN`, `FinykBackup` type,
  `FINYK_FIELD_TO_STORAGE_KEY`, and pure `normalizeFinykBackup` /
  `normalizeFinykSyncPayload` helpers.
- `src/backup.test.ts` — vitest coverage for version gating, empty
  object / non-object rejection, array vs object field validation,
  networthHistory / customCategories / dismissedRecurring malformed
  row rejection, and compact (`b/s/mp`…) vs full backup payloads.

Web side (`apps/web/src/modules/finyk/lib/finykStorage.ts`,
`lib/finykBackup.ts`, `lib/demoData.ts`) becomes a thin storage
adapter that re-exports the pure symbols under their existing names
(`FINYK_STORAGE_KEYS`, `FINYK_BACKUP_VERSION`, `FinykBackup`,
`normalizeFinykBackup`, `normalizeFinykSyncPayload`,
`FINYK_MANUAL_ONLY_KEY`) — no call site in apps/web needs to change.
`readFinykBackupFromStorage` and `persistFinykNormalizedToStorage`
stay on the web side since they are storage-bound.

New package exports: `@sergeant/finyk-domain/storage-keys` and
`@sergeant/finyk-domain/backup`.

Migration doc section 2.4 updated.
